derive: move Parsed into parser module
